syntax: Initial draft of arrays and predicates
This commits adds a draft in the `Syntax.md` document for arrays, which
are primitives as opposed to `List<T>`. Additionally, we introduce
support for the `funcname?` sugar, which turns a function implicitly
into a "predicate" - that is, it implicitly returns a `Bool`. It is best
suited for short, concise functions that can be written with the `=>`
syntax.